+++ Final/lib/Frontend/data_display.h
@@ -4,8 +4,8 @@
 
 #include <stdbool.h>
 
-void draw_graph(int size, float *data, float max_val, float threshold);
-void print_rating(float data, int threshold);
-void print_data(float data[]);
+void draw_graph(int size, double *data, float max_val, float threshold);
+void print_rating(double data, int threshold);
+void print_data(double data[]);
 
 #endif
